add implements adParts and User

+++ src/main/java/com/example/projectmetis/service/ServiceInterface.java
@@ -1,5 +1,6 @@
 package com.example.projectmetis.service;
 
+import com.example.projectmetis.models.User;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -8,8 +9,8 @@ import java.util.List;
 public interface ServiceInterface <T, C >{
     @NotNull List<T> getAll();
     @Nullable T getByName(@NotNull String name);
-    @Nullable T getById(@NotNull Integer id);
-    @Nullable T deleteById(@NotNull Integer id);
+    @Nullable T getById(@NotNull Long id);
+    @Nullable T deleteById(@NotNull Long id);
     @Nullable T deleteByName(@NotNull String name);
     void deleteAll();
     @Nullable T edit(@NotNull C dto);
